hmap_backend/api/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
bcrypt==4.1.2
billiard==4.2.0
boto3==1.34.93
botocore==1.34.93
build==1.2.1
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
cftime==1.6.3
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==42.0.5
decorator==5.1.1
Django==4.2.11
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-guardian==2.4.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-flex-fields==1.0.2
drf-yasg==1.21.7
fabric==3.2.2
geojson==3.1.0
idna==3.7
importlib_metadata==7.1.0
inflection==0.5.1
invoke==2.2.0
jmespath==1.0.1
kombu==5.3.7
MarkupSafe==2.1.5
netCDF4==1.6.5
numpy==1.24.4
packaging==24.0
pandas==2.0.3
paramiko==3.4.0
pip-tools==7.4.1
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
PyNaCl==1.5.0
pyproject_hooks==1.0.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
s3transfer==0.10.1
sentry-sdk==2.0.1
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
wrapt==1.16.0
zipp==3.18.1
